HHVM Debugger: Don't break into __toDebugDisplay during VariablesCommand
Summary:
When evaluating scope objects that have a __toDebugDisplay, we should not break into the __toDebugDisplay method (even if it has a breakpoint). This results in spurrious breaks in some cases, and infinite recursion in the case where there's actually a breakpoint in the method.

Breakpoints in this method may be hit during an evaluation call.

Reviewed By: velocityboy

Differential Revision: D13822067

fbshipit-source-id: 754c693f0117e5c1c7a6fea7753516b0e7ec08c1