Remove Full_fidelity_syntax_tree.from_root
Summary:
The way we handle the file mode in the FFP is mildly clowny: the mode information lives in the parser env, but nothing guarantees that the env that you pass into `Full_fidelity_parser.parse_script` has the same mode written in the file! An example of this is `Full_fidelity_syntax_tree.from_root`. It takes in a parsed tree and an env, and then sets its mode based on the tree, but we don't know if the tree was parsed in the right mode. Most functions take a default env, whose mode is `None`.

In this diff we kill the `from_root` function. This is only the first step—the win in this diff is that we eliminate the `get_mode` function, which is literally a duplicate of `Full_fidelity_parser.parse_mode`. Now there's exactly one place in the code where we parse a file's header.

But, we must still guarantee that the parser uses the mode defined in the file to parse the rest of the file. This suggests that the mode should live on the parser, not on the parser env. We'll do that in a followup diff.

Reviewed By: losvald

Differential Revision: D13871939

fbshipit-source-id: a6bd16f2475db6f748073083bc14b53d28564968