Eagerly solve the type of an unpacked function argument when enforcing it is unpackable
Summary: We do a case analysis on the type of an unpacked argument to enforce it is unpackable; in particular, if the type is a tuple, we check that the types of the components match the types of the function parameters.  In the new inference algorithm, we need to eagerly solve the type to expose its structure to the case analysis.

Reviewed By: andrewjkennedy

Differential Revision: D13895475

fbshipit-source-id: cddfaf6c298068e4625fe873317eb7571e168409