New inference: simplify tyvars plumbing
Summary:
Major refactoring of our tyvars plumbing. Instead of passing around "freshly generated type variables" explicitly, we store sets of type variables on a stack in `env`, pushing an empty set with `Env.open_tyvars` when starting to check an expression (or statement, or other expression form), implicitly adding to the top set on the stack when `fresh_unresolved_type` is called, and popping the set using `Env.close_tyvars` once the expression has been checked. In the meantime, we can use `Env.get_current_tyvars` in order to access the type variables to set the variance, or solve constraints on those type variables.

Quite a lot of plumbing remains; this will be removed in a subsequent diff.

Also, it's likely that we can still improve the factoring, for example we might always solve when popping the type variables.

Reviewed By: manzyuk

Differential Revision: D13882045

fbshipit-source-id: 6f64fb5b1819e142ef4b58a4eceaedf4e78e96c8