Add support for a per repo .hhvmconfig
Summary: Building on the RepoOptions configuration introduced in D13832896 and D13832897 this diff adds support to begin loading a per repo configuration file. Before parsing a new file hhvm will search for a file named `.hhvmconfig.hdf` (similar to Hack's `.hhconfig`) in the directory tree beginning with the file in question and searching upwards. Repo configs are expected to be in HDF form and a cache similar to the unit cache is maintained in a tbb hashmap in `runtime-option.cpp`.

Reviewed By: markw65

Differential Revision: D13832898

fbshipit-source-id: 72ecb2c781c41d9f45fb0e0b8a0747aed478354f