Thread RepoOptions through unit-cache and execution-context
Summary:
In this diff we begin passing the new RepoOptions structure introduced in D13832896 through all of the callsites that will require it. Broadly speaking we want to ensure (1) any files loaded via unit-cache.cpp load RepoOptions prior to cache lookup (and use those options as a part of the cache key), (2) any evaled code can be compiled using RepoOptions loaded for the file containing the eval, and (3) that the packager used for repo builds also loads RepoConfig files.

Note that as with D13832896 this diff should effectively be a noop (but does add the additional check to `isChanged` in unit-cache.cpp for the RepoOptions hash.

Reviewed By: markw65

Differential Revision: D13832897

fbshipit-source-id: 48dde870e2de7867da5689d300f4690682749d1d