Run verifier on hhbbc output
Summary: When running the verifier on hhbbc output the global string table is in write mode. For the most part it was possible to tweak existing checks to work in this mode by extending LitstrTable::contains to work in both modes, however, checks for `InitProp` and `CheckProp` to validate property names had to be weakened in this mode.

Reviewed By: oulgen

Differential Revision: D13867233

fbshipit-source-id: 25f0656e0d54024de3cbbcee33d510c0c3274511