Migrate `fun_` to naming to operate on the aast
Summary:
We want to remove the "legacy" AST and unify it with the AAST. To do this, we need to convert the Naming module, which currently operates over the legacy AST, to be able to use the AAST.

This diff adds `aast_fun_` as an option to work on the `Aast.fun_` as a copy of `fun_` which works on `Ast.fun_`.

To do this, there are many parts of `Aast.fun_` that still reference the `Ast` such as `func_unnamed_body` holding an `Ast.block` that need to be changed. Change `func_unnamed_body` to use `Aast.block` and `Aast.param`. Add `aast_type_constraint` and use it in `genv`

Then alias `fun_` with `Ast_to_nast.on_fun |> aast_fun_`.

Reviewed By: arun11111

Differential Revision: D13696962

fbshipit-source-id: d56ce1dc95565bd04abf911800adec9b2096b77e