Functorize Aast over FuncBodyAnnotations, In the middle...
Summary:
We want to remove the notion of `NamedBody` and `UnnamedBody` from the `aast`. We functorize the extra bits (besides the `block` body) out in `aast` so that `nast` and `tast` will have to pass in their own versions.

After that is done, this diff also converts all callsites that previously used the original struct.

Reviewed By: Wilfred

Differential Revision: D13866650

fbshipit-source-id: 30095ede29c5a937a013efdeff4b95e9b1b1aea2