[AArch64] Correctly handle smashables which have been optimized in place.
Previously, there was special handling in bindJmp() (tc-bind.cpp) if
the veneer for a smashable jcc was already smashed and optimized from
a LDR, BR sequence to a simple B. This optimization is done in
optimizeSmashedJcc() (smashable-instr-arm.cpp). The special handling
involves recognizing that a smashed and optimized jcc no longer has
a smashable target as returned by smashableJccTarget(), and instead
a null ptr is returned.

This handling of a null ptr for the smashableXxTarget() was extended
to smashable jmps and calls. While these two smashables don't have
explicit optimizeSmashedXx() phases, they can be optimized while
smashing directly. Smashable jmps, jccs, and calls can all be
optimized during smashing and the result is that the veneer is
bypassed completely, which makes the respective smashableXXTarget()
calls all return a null ptr as in the scenario above for optimizing
a jcc after smashing. Only debug builds catch this case due to an
assert.