hackfmt: specify options via .hhconfig
This change allows users to control hackfmt options through .hhconfig.
The following options can be configured:

- hackfmt.add_trailing_commas
- hackfmt.indent_width
- hackfmt.tabs
- hackfmt.line_width

Each of the options will be superseded by the equivalent command-line
arguments if provided, except for add_trailing_commas, which is not
exposed via CLI args.

hackfmt will make a best-effort attempt to discover the location
of .hhconfig. It will search ancestor paths of the given file, or
if no file is provided, the working directory. If --root is
provided explicitly, hackfmt will look for .hhconfig under that
directory.

Addresses https://github.com/facebook/hhvm/issues/8266