allow requests to catch signals (such as SIGTERM) in server mode
Summary:
Use a dedicated thread to receive and handle certain signals (particularly ones that terminates the process) in a synchronous way.
CLI mode has been running with this mechanism.  This diff extends it to the server mode.
When a signal is received by the server process, it will broadcast SIGTERM (instead of the original signal) to all ongoing requests.  If a request registers a handler, it gets a chance to run cleanup code.
Beware that if the request doesn't finish cleanup within the shutdown grace period, the process will exit abruptly anyway.  So there isn't a real guarantee that the cleanup will happen, unless we have very big grace period.

With this mechanism, we also unify the server shutdown process: `HttpServer::stopOnSignal()` sends the signal to the handler thread, which calls `HttpServer::stop()`.

Reviewed By: markw65

Differential Revision: D13925122

fbshipit-source-id: 73bfa62703b5910ac57ad9e35bf6a472f2d2d27e