Eagerly solve the types of the arguments of await expressions
Summary:
Yet another case analysis that doesn't deal with type variables correctly and that benefits from eager solving.

Here, given the type of an expression `e`, we are inferring the type of the expression `await e`.  The problem is that the `await` operator has an overloaded type.  If the type of `e` is `?Awaitable<T>`, then `await e` has the type `?T`.  However, if we know that `e` is non-nullalbe and has type `Awaitable<T>`, we can infer a more precise type `T` for `await e`.  This logic is hard to express at the moment using only subtyping (it could be expressed using an intersection type for `await`), so we are forced to perform a case analysis on the type of `e`.

Reviewed By: andrewjkennedy

Differential Revision: D13919745

fbshipit-source-id: 57ff6490799c303b9058cfff710661c39577d661