simplify singleton Tunresolved
Summary:
After performing a union, if the result is a singleton Tunresolved [t], we simplify to t.

We only do that with new_inference because the old inference treats `Tunresolved [t]` and `t` differently: the first is allowed to "grow" while the second isn't: a `$v` passed as argument will have type `vec<int>` and it's not possible to add a string in it, while a `$v` initialized as `vec[0]` will have type `vec<Tunresolved[int]>` and one can put a string in it.

Reviewed By: andrewjkennedy

Differential Revision: D13919169

fbshipit-source-id: eb63d84b037dda6d36ec19504f3381114abbff78