New inference: set tyvar variance to account for lambda arguments to function application
Summary:
When checking function applications to generic functions such as
```
function foo<T1,...,Tn>(t1,...,tn):u
```
where some of the `ti` are function types, and the actual arguments are lambdas, we need to be aware that bidirectional checking of the lambdas will assign parameters types possibly involving the `Ti` generic parameters. This means that we need to take account of more than just the appearance of type parameters in `u` when setting the variance, as checking of the bodies of the lambdas needs to be aware of those type parameters. For example, consider
```
function foo<T>(T $x, (function(T):void $f):Contra<T>
$r = foo($a, $z ==> $z->bar());
```
Our default setting of variance would treat the fresh type variable (say `#1`) generated for `T` as contravariant. We would generate some constraint when checking the first argument, getting
```
C <: -#1
```
But when we check the body of the lambda, we would eagerly solve the type variable to get `#1 := mixed`, even though we have in our hand the variable `$x` of type `#1` i.e. the type variable appears *covariantly* in the type of an expression that we are depending on.

The solution in this diff is to set variance also taking into account *inferred* types of parameters to lambda arguments. This is a bit clunky - we are replicating work that is done later when doing bidirectional checking of the lambda, in particular dealing with unions and nullables. Currently considering alternative approaches.

Reviewed By: CatherineGasnier

Differential Revision: D13878039

fbshipit-source-id: c7848e607d897016034b6465ff8f21313ebea01b