Templatize DualDispatch related utilities: Commute and DDHelperFn
Summary:
The goal of this diff is to make it easier to add a new DataTag to HHBBC's type system. Currently, each new DataTag requires adding quadratic number of new function to explain the relation between the new DataTag and the old ones even though, the new one might not need to interact with all of the old ones.

This diff adds `DataTagTrait` to `Commute` and `DDHelperFn` to explain which tags need to interact with each other.

The motivation for this diff is to make the next diff simpler.

Reviewed By: markw65

Differential Revision: D13935476

fbshipit-source-id: fd259d036a85486d3099584678167817fdbc0100