HHVM Debugger tests: don't poll child script process exit
Summary: The test harness is polling waiting for the child HHVM instance that runs a debugged test script to actually exit. On some builds, this can take longer than the test was allowing, causing a timeout and the test to fail. This diff closes the process handle but does not poll exit status.

Reviewed By: alexeyt, markw65

Differential Revision: D13950290

fbshipit-source-id: 091a93b4b7a7e65bcf22540038df96da86da764c