Report error for *all* incorrect uses of variant type parameters
Summary:
Hack's current approach to checking consistency of uses of type parameters wrt their declared variance is as follows: *infer* a variance for each type parameter, using their uses, and then check this inferred variance against the declared variance. Unfortunately this means that if there is an inconsistent use (e.g. a covariant type parameter is used as the type of a parameter to a function), then only *one* such bad use is reported.

(In particular, this makes it hard to fix errors with variance at the use sites, as when one usage error is fixed, another one can be "revealed").

Better to simply track the variance of uses, and for each use check that the declared variance is consistent.

Reviewed By: Wilfred

Differential Revision: D13957810

fbshipit-source-id: b402c0a1944d6ce2e4f285e85bd13c5fcb267586