New inference: maintain position information in type variable environment
Summary:
When we report a "circularity" error (trying to solve a type variable v with a type t that contains v) we have no position information, so the error that is surfaced is useless.

Fix this by maintaining a position in the type variable environment.

Also do a little refactoring to save repeated lookups of a variable in the environment.

Reviewed By: CatherineGasnier

Differential Revision: D13972797

fbshipit-source-id: d569661aae21b54e9293cd7ddd609255a83a2366