Enable safe instanceof in partial mode
Summary: If `C` is a generic type of arity `n`, then for the `C` in `$x instanceof C` we will generate a type `C<v1, ..., vn>`, where `v1`, ..., `vn` are fresh type variables.  In strict mode (with `safe_instanceof` enabled), these will be dropped, and replaced by fresh *type parameters*.  But in non-strict mode, they will remain, to be unsafely unified (in legacy type checking) with whatever comes along inside the branch of the conditional.  It's time to get rid of this and enable `safe_instanceof` in non-strict mode.

Reviewed By: andrewjkennedy

Differential Revision: D13987035

fbshipit-source-id: 40106e4cffee9e0972748df2397c8bafcf68d5ad