Pass through options to GlobalOptions.make in ServerConfig
Summary: In this diff, we update the usage of `GlobalOptions.make` in ServerConfig so that we don't duplicate the defaults. This already caught a few instances of the defaults falling out of sync. In the future, this will no longer be the case—defaults will **only** be defined in GlobalOptions.

Reviewed By: andrewjkennedy

Differential Revision: D13980946

fbshipit-source-id: 77eb4b664c858994b79f00de95ceff071879e085