Make all arguments to GlobalOptions.make optional
Summary:
This function is bad and we should feel bad. Because all of the arguments are required, we need to write out the entire function wherever we want to flip a few options, e.g. in integration tests. Plus, it forces us to duplicate defaults in ServerConfig and hh_single_type_check, which means they can get out of sync. If a user wants the compiler to guarantee they set every key, they can write out a record manually.

In this diff, we make all of the arguments optional, allowing us to remove them wherever they line up with the defaults. In the next diff, we'll update the usage in ServerConfig so that the defaults are defined in exactly one place.

Reviewed By: andrewjkennedy

Differential Revision: D13979916

fbshipit-source-id: 5b23136f106f2e3c729a8d3bd724497cfcf408e0