Pass through options to GlobalOptions.make in hh_single_type_check
Summary:
In this diff, we update the usage of `GlobalOptions.make` in hh_single_type_check so that we don't duplicate the defaults, so that defaults are **only** in GlobalOptions.

In order to prevent tests from beginning to fail, I left some defaults out of sync with `GlobalOptions.default`. In followup diffs, we can investigate fixing the tests so that they reflect the new defaults, or have the proper options set to test the particular behavior.

Reviewed By: andrewjkennedy

Differential Revision: D13981806

fbshipit-source-id: 0282d00e2a03b3cadf055ba4d2542fffe8537e67