Remove incorrect reified generics clsref irgen optimization
Summary:
As markw65 pointed out, this optimization is incorrect since it might raise an error while jitting code. The jitting thread may not even be associated with a request thread. We can instead opt to emit raise error ir instruction but if the preceding bytecode was ReifiedName, it might add the entry for name (but it hasn't executed yet, because we're only jitting the tracelet, not executing it).

We could get around this restriction by emitting yet another ir instruction to get the generic but then we are back to the beginning and we wouldn't get enough for all the complexity. Lets just kill it.

Reviewed By: markw65

Differential Revision: D13976274

fbshipit-source-id: 5d2a8d0246c13f9f016755244f5b3e2453ea81d5