Combine closing type variables with solving
Summary: Andrew and I were debugging a failing test case, and Andrew suggested that we should solve the current set of type variables *after* popping it from the stack.  (Currently, we get the current set of type variables, solve it, and then pop it from the stack.)  The reason is that solving can introduce fresh type variables through freshening (which we currently don't track but should) and those should be added to the parent set of type variables, i.e., they should be solved when type-checking the enclosing expression.  It turned out later that the failing test case started to pass once I rebased on top of Andrew's D13878039 and the changes from this diff were not necessary, but Andrew asked me to publish this diff anyway.  There are no tests that are improved by this change though.

Reviewed By: andrewjkennedy

Differential Revision: D13958364

fbshipit-source-id: 3e2359f121e22b9e9e5548ac007f29458b890c21