deal with type constants
Summary:
When we localize a function type with where constraints, we might localize type const access like T1::T. Whenever we localize type consts like T1::T, create a tvar v2 representing T1::T, and add it in the node for v1, representing T1:
```
v1 -> {
  upper_bound: [ StringBox; ... ]
  lower_bounds: ...
  appears_covariantly: ...
  appears_contravariantly: ...
  type_consts: varid SidMap.t = [
    "T" -> v2
  ]
}
```
Note: if we localize T1::T::T, then we would need v1 for T1, v2 for v1::T and v3 for v2::T.

For all present and future upper bounds U of #1, localize U::T as a type variable #3 with the same constraints as T in U, then make #2 equivalent to #3.

For all present and future lower bounds L of #1, simply make #2 equivalent to L::T.

For upper bounds, we're just checking that the constraints on U::T are satisfied. For lower bounds L however, we make #1::T equal to the type constant in L, possibly making it an expression dependent type if T is abstract in L.

Reviewed By: andrewjkennedy

Differential Revision: D13958402

fbshipit-source-id: 23b8d157fd78ae27d777904bbea53a166f6a939b