minor fix in typing_union: normalize unions with null
Summary:
Normalize unions like (null | t1 | t2) as ?(t1 | t2) in Typing_union.
I couldn't come up with a test for which that matters, in the current code base, but with another diff of mine, it would avoid failing to determine that ?t is indeed a subtype of (null | t).

Reviewed By: andrewjkennedy

Differential Revision: D14004284

fbshipit-source-id: 31f05348cb34a13a30d472934772c456a80354b9