Emit StMem's instead of InitObjProps in HHIR for objects with small number of properties
Summary:
We were emitting the property initializations inline for objects with a small
number of properties at Vasm level, when lowering InitObjProps.  However, this
was too late to benefit from the optimizations performed at HHIR level.  Doing
this when emitting HHIR allows both dead stores and DecRefs to be eliminated in
HHIR when the following constructor call is inlined.

Reviewed By: paulbiss

Differential Revision: D13995838

fbshipit-source-id: 9faf2a5a246210e029333dd624cfcaf5f8fce09b