Remove emit_call_lhs_with_this
Summary:
emit_call_lhs_with_this was a complicated and buggy way to implement ClsRefGetC
of `$this`. Complicated because it tries to preserve correctness in the
presence of changing `$this`, which won't happen in these contexts (and won't
happen eventually at all when we close all the loopholes). Buggy because the
`instrs` instruction sequence may be using the same unnamed local that the
function allocates internally.

Reviewed By: oulgen

Differential Revision: D14018550

fbshipit-source-id: f04908f6dca73c7717927cacd4cb3981bdd399cc