Delay error reporting in lowering so that CST errors can be migrated
Summary:
Currently, the lowering (CST->AST) is effectful in a sense that errors are
either raised or accumulated by mutating data in Errors module.  However,
that doesn't play well with our goal of moving pre-lowering errors
from CST to (A)AST, so that we can get rid of CST and do direct lowering.

Because errors emitted during lowering are less fine-grained / more confusing,
they are currently prioritized, both by:
- design: FFP error checks (full_fidelity_parser_errors.ml) runs *before* lower
- ad-hoc: in lowering errors there is a check to emit only if no FFP error

This diff separate effects from result of the lowering as follows:
- in raise_parsing_error (entry point to raising/accumulating errors in lower):
  record the position and message of the error and add to a shared list via ref
- make effects later (after a new phase on AST) via: `flush_parsing_errors`
  which goes over the list, and replays the raise/accumulate behavior that
  previously existed in `raise_parsing_error`
  (note that we could not fully replay if lowering ever handles those exceptions;
  fortunately, this is not the case and should stay that way)

Reviewed By: oulgen

Differential Revision: D13965774

fbshipit-source-id: be5ef30320a8b63cc0083e7871ba8a6bf4b5b04f