Run parsing, Naming, and Decl using global ParserOptions and TypecheckerOptions
Summary:
Today, hh_server has many internal APIs with a shape like this:

```
Parser_heap.get_from_parser_heap : ParserOptions.t -> Relative_path.t -> Ast.program
Naming.class_              : TypecheckerOptions.t -> Ast.class_ -> Nast.class_
Shallow_classes_heap.get   : TypecheckerOptions.t -> string -> shallow_class option
Decl.declare_class_in_file : TypecheckerOptions.t -> Relative_path.t -> string -> unit
Typing_lazy_heap.get_class : TypecheckerOptions.t -> string -> class_type option
```

All of these APIs are backed by worker-local or sharedmem caches. From the type signature, one might reasonably conclude that the contract of a function like `get_from_parser_heap` is as follows: "parse the file at the given path using the given parser options and return the resulting AST." Knowing that the API was backed by a sharedmem cache, one might conclude instead that the contract is: "If the file at the given path has already been parsed with the given parser options, return the stored AST. Otherwise, parse it using the given parser options and return the result."

Unfortunately, this is not true. The parser and typechecker options are not part of the key type for any of these caches. We don't want them to be, either--we are memory-constrained enough as it is without the possibility of storing multiple parse trees for each file.

As a result, when a consumer of `get_from_parser_heap` requests an AST parsed with a given set of parser options, they may get it, or they may get an AST parsed with a totally unrelated set of options. With `Shallow_classes_heap`, `Decl`, and `Typing_lazy_heap`, the situation is even worse--since these modules fetch ASTs from `Parser_heap` before declaring them, the given options may be used for declaring, but not for parsing (since the typechecker options are fed directly into the parser--unintuitively, ParserOptions.t and TypecheckerOptions.t are aliases for the same type). The given options could also be used for both declaration and parsing, or not at all, or for parsing but not declaration.

These semantics do not seem useful, so in lieu of accepting options as a part of these APIs, this diff proposes that we set `GlobalParserOptions` and `GlobalNamingOptions` when the server starts, and use those global options for all of these cache-populating operations for the entire lifetime of the server.

As far as I can tell, we do not have any real-world use cases where it is necessary to change the options used for parsing, naming, or decl during the lifetime of a Hack server. It looks like the only option read by parsing, naming, or decl which is actually changed while a server is running is `tco_assume_php` (always via `GlobalOptions.make_permissive`). This option seem to be changed most frequently by IDE and build tools, seemingly as an error suppression mechanism (possibly because it predates `Errors.ignore_`, which was added in D1571225). Since we now have `Errors.ignore_`, this doesn't seem necessary (and indeed, our IDE test cases pass after this diff, which removes `make_permissive`).

Reviewed By: ljw1004

Differential Revision: D13551879

fbshipit-source-id: 0102c3e8c96ff2548ec02909aba9bdde460074aa