Use Shallow_classes_heap in Decl
Summary: As we start to use shallow classes as more than just a precursor to Decl, we will want to avoid redoing unnecessary work by holding on to a worker-local cache of shallow classes. We already have such a cache in Shallow_classes_heap--we just need to make use of it in Decl.

Reviewed By: vassilmladenov

Differential Revision: D13458772

fbshipit-source-id: 001aac0c62ed6e3ba91f5619e1e7fe6fb64ca3ed