Flag to disable UNSAFE_EXPR
Summary:
UNSAFE_EXPR disables the typechecker for certain regions of code, meaning we can't actually be sure that the typechecker is catching everything. We would like to change instances to HH_FIXME (or equivalent) where the typechecker still catches everything, even if we then ignore it.

This change adds a flag to treat UNSAFE_EXPR like a normal comment so we (and OSS users) can get a feel for what its removal would break.

Reviewed By: kmeht

Differential Revision: D13997015

fbshipit-source-id: e6335ad7c614f735854e0a3255beb0a895fae1bc