Add nullable smart constructor to Typing_make_type
Summary: Add a `nullable` smart constructor to `MakeType`. This avoids trivial cases of "double nullables" by looking at the unexpanded type that is passed in. A more sophisticated approach would need an environment, in order to expand type variables; and in any case, we will probably just replace all of this by union with `null` in future.

Reviewed By: CatherineGasnier

Differential Revision: D14041981

fbshipit-source-id: c75b247c44bcd90935aa296b2ec5d9bb5581ccc8