New inference: clean up variance setting
Summary:
The API for variance setting is always used in a fixed pattern; capture that in the API instead.

Also, fixed an issue with setting of variance for expressions such as `new C(...$x ==> e...)` where the variance according to the created object isn't set until *after* the lambda argument is checked, in contrast to the situation for ordinary calls. (The problem is that the function type for the constructor doesn't include the type of the created object as its return type.) No existing tests surface this issue. It was picked up when tweaking the use of eager solving of type variables in a forthcoming diff.

Reviewed By: CatherineGasnier

Differential Revision: D14064348

fbshipit-source-id: b78ebe3fd00602ffe8e5dfdb71193522bfd35d38