Default define() to a warning
Summary:
We've had define as a fatal in production since December 7th, D13324047. This changes the `DisableDefine` option to a warning by default.

This change causes the warning to be thrown when tests are run. I've updated the `expect` files to expect the warning for tests that are explicitly testing the behaviour of `define`.

In other tests, that use `define` but are not testing it's behaviour I've removed those calls and replaced them with `consts` where possible.

Reviewed By: kmeht

Differential Revision: D13427189

fbshipit-source-id: a78dc4c6a20d9b3a77d366e0da2a2592596717c9