Add soft type hints to reified generics
Summary:
This diff introduces new mechanisms to reified generics parameter and return type verification to enable simpler migration from erased to reified generics, namely it introduces soft type hints.

1. Soft at Call site

```
function f(C<int> $x) {}
```
function `f`'s first parameter, `$x`, is checked whether it is a `C<int>` and if it is not `C` we raise error but if it is `C` and not `int`, we only raise a warning.

2. Warn at Declaration Site

```
class C<<<__Warn>> reify T> {}

function f(C<int> $x) {}
```
Similar to soft at the call site, this time soft type hint applies to every single time this generic is used.

Benefits of these two approaches is that it enables gradual migration.

Reviewed By: markw65

Differential Revision: D14056294

fbshipit-source-id: 52fcbb941c858a83a2c660f3166650c8ea478124