Move accessing a reified generic bytecode into the emitter and add HHBBC optimizations
Summary:
We no longer need a specific bytecode to access a `ReifiedGeneric` since it is stored as the first non parameter local for functions/methods and as a property of the class for class level generics.

Moving it to the emitter enables us to take advantage of varray/vec access optimizations already built in to the JIT.

In addition to this, this diff adds some optimization to HHBBC since we know that the container for reified generics (property or local) is a varray/vec with a known size. So, we can set its type to known size DataTag.

Reviewed By: markw65

Differential Revision: D13954820

fbshipit-source-id: f63c8d0d8db5a5fb543273dbc8bfbedc4e42c62f